Add grpc handlers logic
